Cleanup the releasenote testing via configuration loading
